pyproject.toml
[project]
name = "dagster_jaffle_shop"
requires-python = ">=3.9,<3.13"
version = "0.1.0"
dependencies = [
    "dagster",
    "dagster-azure>=0.26.14",
    "dagster-dbt>=0.26.14",
    "dagster-dg>=0.26.14",
    "dagster-duckdb>=0.26.14",
    "dagster-duckdb-pandas>=0.26.14",
    "dbt-duckdb>=1.9.3",
    "duckdb>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_jaffle_shop"
